fix(kpi #1): PSY 분모를 평균 사육 모돈 재고로 정정(스펙 §1)
수치감사(라이브 실증): calculate_psy가 읽던 v_farm_psy 뷰는 분모를
COUNT(DISTINCT 그 해 이유한 모돈)으로 계산 — 이유 안 한 모돈은 NULL 버킷으로 빠지고
분모가 사육두수가 아닌 이유실적 모돈수라 PSY 심각 과대(20두 농장 9/1=9.0 관측).
스펙 §1대로 분모=연도 12개월 각 월초 활성 모돈 재고(입식~퇴출 윈도우, 소프트삭제 반영)
평균으로 인라인 재구현(버그 뷰 사용 중단). 분자=연간 SUM(weaned_count).
엣지: 이유 0건 → PSY=0(NULL 아님), 활성 0두 → PSY=NULL. avg_sow_count int→float.

테스트: 이유 실적 1두/합계 200 + 활성 10두 → PSY=20.0(옛 뷰라면 200). 회귀 792→795.

+++ api/app/schemas/kpi.py
@@ -62,7 +62,7 @@ class _KpiValueInternal(BaseModel):
 class PsyDetail(BaseModel):
     farm_id: UUID
     year: int
-    avg_sow_count: int
+    avg_sow_count: float  # 월별 활성 모돈 재고 평균(소수 가능) — 스펙 §1
     total_weaned: int
     psy: float | None
     benchmark_avg: float | None
